Changed buf_len to size_t type, added const qualifiers

+++ src/resource_records.h
@@ -25,7 +25,7 @@ int form_question_record(unsigned char *buf, ssize_t buf_size,
             uint16_t id, const char *hostname, int type);
 
 int parse_dns_records(unsigned char *buf,
-            int buf_len, const char *hostname, uint16_t id, dns_rr **out);
+            size_t buf_len, const char *hostname, uint16_t id, dns_rr **out);
 
 int has_expired_records(dns_rr *records);
